Added option (useful for testing) to not print png files for simulation, just the txt files.  This can be done when say the time of a phenomenon is known, and multiple reruns need to be performed to observe different quantities without the extra overhead (and text in output) from png and gif formation.

+++ simulate.sh
@@ -1,6 +1,14 @@
 #!/bin/bash
+count_pngs=`ls -1 simulations/*.png 2>/dev/null | wc -l`
+if [ $count_pngs != 0 ]
+then
 rm ./simulations/*png
+fi
+count_txts=`ls -1 simulations/*.txt 2>/dev/null | wc -l`
+if [ $count_txts != 0 ]
+then
 rm ./simulations/*txt
+fi
 python ./src/main.py $1 $2 $3 $4
 declare pids
 # for pid in $(pgrep 'draw'); do
@@ -18,4 +26,10 @@ declare pids
 # done
 # echo $end
 # ./draw $start $2
-yes | ffmpeg -i simulations/plane_%03d.png simulations/simul.gif
+# if [ "$(ls -A simulations/plane_*png)" ]; then
+#if ls ${simulations}/*.png;
+count_pngs=`ls -1 simulations/*.png 2>/dev/null | wc -l`
+if [ $count_pngs != 0 ]
+then
+  yes | ffmpeg -i simulations/plane_%03d.png simulations/simul.gif
+fi
